+++ include/trace/events/kvm.h
@@ -296,14 +296,14 @@ TRACE_EVENT(kvm_halt_poll_ns,
 	trace_kvm_halt_poll_ns(false, vcpu_id, new, old)
 
 TRACE_EVENT(kvm_dirty_ring_push,
-	TP_PROTO(struct kvm_dirty_ring *ring, u32 slot, u64 offset),
+	TP_PROTO(struct kvm_dirty_ring *ring, u64 slot, u64 offset),
 	TP_ARGS(ring, slot, offset),
 
 	TP_STRUCT__entry(
 		__field(int, index)
 		__field(u32, dirty_index)
 		__field(u32, reset_index)
-		__field(u32, slot)
+		__field(u64, slot)
 		__field(u64, offset)
 	),
 
@@ -316,7 +316,7 @@ TRACE_EVENT(kvm_dirty_ring_push,
 	),
 
 	TP_printk("ring %d: dirty 0x%x reset 0x%x "
-		  "slot %u offset 0x%llx (used %u)",
+		  "slot %llu offset 0x%llx (used %u)",
 		  __entry->index, __entry->dirty_index,
 		  __entry->reset_index,  __entry->slot, __entry->offset,
 		  __entry->dirty_index - __entry->reset_index)
